requirements.txt
streamlit==1.18.1
requests==2.28.1
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
groq==0.22.0  # Updated to the latest version
